// src/app/model/rejectionreason.ts
export interface RejectionReason {
    id: string;
    code: string;
    createDate: string;
    nrlcode: string;
    qctocode: string;
    setmiscode: string;
    description: string;
    forprocess: string;
    howto: string;
    smscode: string;
    forcrm: string;
    formanager: string;
    forseniormanager: string;
    softdelete: string;
}